fix duplicate code
